Section sessions by start time
The current implementation is rather simple and only display a section header if the item’s position is 0 or if its start time is different from the one at position - 1.
A possible enhancement regarding UX would be to have sticky headers so that start time is always visible. Unfortunately, I haven’t found a simple library to do this…